+++ x/pagination.go
@@ -18,8 +18,8 @@ func ParsePagination(r *http.Request) (page, itemsPerPage int) {
 	return migrationpagination.NewDefaultPaginator().ParsePagination(r)
 }
 
-func PaginationHeader(w http.ResponseWriter, u *url.URL, total int64, page, itemsPerPage int) {
-	migrationpagination.PaginationHeader(w, u, total, page, itemsPerPage)
+func PaginationHeader(w http.ResponseWriter, u url.URL, total int64, page, itemsPerPage int) {
+	migrationpagination.PaginationHeader(w, &u, total, page, itemsPerPage)
 }
 
 type Page struct {